Remove MSVC14 warnings

+++ src/aes.c
@@ -239,7 +239,7 @@ int aes_encrypt_cbc(const BYTE in[], size_t in_len, BYTE out[], const WORD key[]
 	if (in_len % AES_BLOCK_SIZE != 0)
 		return(FALSE);
 
-	blocks = in_len / AES_BLOCK_SIZE;
+	blocks = (int)(in_len / AES_BLOCK_SIZE);
 
 	memcpy(iv_buf, iv, AES_BLOCK_SIZE);
 
@@ -262,7 +262,7 @@ int aes_encrypt_cbc_mac(const BYTE in[], size_t in_len, BYTE out[], const WORD k
 	if (in_len % AES_BLOCK_SIZE != 0)
 		return(FALSE);
 
-	blocks = in_len / AES_BLOCK_SIZE;
+	blocks = (int)(in_len / AES_BLOCK_SIZE);
 
 	memcpy(iv_buf, iv, AES_BLOCK_SIZE);
 
@@ -287,7 +287,7 @@ int aes_decrypt_cbc(const BYTE in[], size_t in_len, BYTE out[], const WORD key[]
 	if (in_len % AES_BLOCK_SIZE != 0)
 		return(FALSE);
 
-	blocks = in_len / AES_BLOCK_SIZE;
+	blocks = (int)(in_len / AES_BLOCK_SIZE);
 
 	memcpy(iv_buf, iv, AES_BLOCK_SIZE);
